Refactored Packet.cs into separate classes under Dto\Packets
