+++ keyforge-network-dotnet/Dto/Packets/ExitPacket.cs
@@ -1,0 +1,10 @@
+﻿namespace KeyforgeNetwork.Dto.Packets
+{
+    public class ExitPacket : Packet
+    {
+        public ExitPacket()
+        {
+            Type = PacketType.Exit;
+        }
+    }
+}
