+++ keyforge-network-dotnet/Dto/Packets/JoinLobbyRequestPacket.cs
@@ -1,0 +1,10 @@
+﻿namespace KeyforgeNetwork.Dto.Packets
+{
+    public class JoinLobbyRequestPacket : Packet
+    {
+        public JoinLobbyRequestPacket()
+        {
+            Type = PacketType.JoinLobbyRequest;
+        }
+    }
+}
