+++ keyforge-network-dotnet/Dto/Packets/LobbyListRequestPacket.cs
@@ -1,0 +1,10 @@
+﻿namespace KeyforgeNetwork.Dto.Packets
+{
+    public class LobbyListRequestPacket : Packet
+    {
+        public LobbyListRequestPacket()
+        {
+            Type = PacketType.LobbyListRequest;
+        }
+    }
+}
