+++ keyforge-network-dotnet/Dto/Packets/LoginRequestPacket.cs
@@ -1,0 +1,21 @@
+﻿using Newtonsoft.Json;
+
+namespace KeyforgeNetwork.Dto.Packets
+{
+    public class LoginRequestPacket : Packet
+    {
+        [JsonProperty(PropertyName = "name")]
+        public string Name { get; set; }
+
+        [JsonProperty(PropertyName = "id")]
+        public string ID { get; set; }
+
+        [JsonProperty(PropertyName = "token")]
+        public string Token { get; set; }
+
+        public LoginRequestPacket()
+        {
+            Type = PacketType.LoginRequest;
+        }
+    }
+}
