+++ keyforge-network-dotnet/Dto/Packets/LoginResponsePacket.cs
@@ -1,0 +1,10 @@
+﻿namespace KeyforgeNetwork.Dto.Packets
+{
+    public class LoginResponsePacket : Packet
+    {
+        public LoginResponsePacket()
+        {
+            Type = PacketType.LoginResponse;
+        }
+    }
+}
