+++ keyforge-network-dotnet/Dto/Packets/PlayerListRequestPacket.cs
@@ -1,0 +1,10 @@
+﻿namespace KeyforgeNetwork.Dto.Packets
+{
+    public class PlayerListRequestPacket : Packet
+    {
+        public PlayerListRequestPacket()
+        {
+            Type = PacketType.PlayerListRequest;
+        }
+    }
+}
