+++ keyforge-network-dotnet/Dto/Packets/VersionPacket.cs
@@ -1,0 +1,15 @@
+﻿using Newtonsoft.Json;
+
+namespace KeyforgeNetwork.Dto.Packets
+{
+    public class VersionPacket : Packet
+    {
+        [JsonProperty(PropertyName = "version")]
+        public float Version { get; set; }
+
+        public VersionPacket()
+        {
+            Type = PacketType.VersionRequest;
+        }
+    }
+}
