+++ keyforge-network-dotnet/Models/PlayerListEntry.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace KeyforgeNetwork.Models
+{
+    public class PlayerListEntry
+    {
+        [JsonProperty(PropertyName = "id")]
+        public string Id { get; set; }
+
+        [JsonProperty(PropertyName = "name")]
+        public string Name { get; set; }
+    }
+}
